[test] refactor
